+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -5,6 +5,11 @@
 * Fix native AoT warnings in `OpenTelemetry.Exporter.OpenTelemetryProtocol`.
   ([#5520](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5520))
 
+* `User-Agent` header changed from
+  `OTel-OTLP-Exporter-Dotnet-{NuGet Package Version}+{Commit Hash}`
+  to `OTel-OTLP-Exporter-Dotnet-{NuGet Package Version}`.
+  ([#5528](https://github.com/open-telemetry/opentelemetry-dotnet/pull/5528))
+
 ## 1.8.0
 
 Released 2024-Apr-02